Prefer the global scope insted of nsURI in application configuration
Trying to find a resource by nsURI first sometimes lead to a situation
where a spurious resource was added to the ResourceSet which didn't
contain the requested object. Thus, linking failed.

By first looking up the imported name in the global scope and falling
back to the nsURI only later, we can ensure that linking always succeeds
if the referenced object is already in the Xtext index.